// src/app/api-client/model/familyPricingResponseDTO.ts
/**
 * TeachingBACKEND API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { BillingInterval } from './billingInterval';
import { PackageTier } from './packageTier';


export interface FamilyPricingResponseDTO { 
    packageId?: string;
    name?: string | null;
    tier?: PackageTier;
    billingInterval?: BillingInterval;
    basePrice?: number;
    pricePerMember?: number;
    familyMembers?: number;
    additionalMembers?: number;
    additionalCost?: number;
    totalPrice?: number;
    maxMembers?: number;
    minMembers?: number;
    breakdown?: string | null;
    basePriceFormatted?: string | null;
    pricePerMemberFormatted?: string | null;
    additionalCostFormatted?: string | null;
    totalPriceFormatted?: string | null;
}
export namespace FamilyPricingResponseDTO {
}